added cool rock bullets and combined player HUD into 1 file
