+++ source/core/src/main/com/deco2800/game/areas/Level2.java
@@ -30,6 +30,7 @@ public class Level2 extends GameArea {
   private static final GridPoint2 PLAYER_SPAWN = new GridPoint2(10, 10);
   private static final float WALL_WIDTH = 0.1f;
   private static final String[] forestTextures = {
+    "images/playeritems/shootingammo.png",
     "images/Player_Sprite/front01.png",
     "images/player_placeholders/PROJECTILE.png",
     "images/obstacle_sprite/cobweb.png",
