+++ source/core/src/main/com/deco2800/game/entities/factories/PlayerFactory.java
@@ -65,7 +65,6 @@ public class PlayerFactory {
                     .addComponent(new PlayerRangeAttackComponent())
                     .addComponent(new DisposingComponent())
                     .addComponent(new PlayerStatsDisplay())
-                    .addComponent(new PlayerInventoryDisplay())
                     .addComponent(animator)
                     .addComponent(new PlayerAnimationController())
                     .addComponent(new PlayerHudAnimationController())
